requirements.txt
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
certifi==2017.7.27.1
chardet==3.0.4
constantly==15.1.0
Django==1.8
django-socket-server==0.0.4
djangorestframework==3.6.4
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
MySQL-python==1.2.5
pytz==2017.2
requests==2.18.4
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
whitenoise==3.3.1
zope.interface==4.4.3
